feat: update and refactor endorsement chain UI
