+++ src/components/EndorsementChain/EndorsementChainContainer.tsx
@@ -1,6 +1,6 @@
 import React, { FunctionComponent } from "react";
 import { useEndorsementChain } from "../../common/hooks/useEndorsementChain";
-import { EndorsementChainLayout } from "./EndorsementChainLayout/EndorsementChainLayout";
+import { EndorsementChainLayout } from "./EndorsementChainLayout";
 
 interface EndorsementChainContainer {
   tokenRegistry: string;
