--- src/components/EndorsementChain/EndorsementChainLayout/EndorsementChain/index.tsx
@@ -1,1 +1,0 @@
-export * from "./EndorsementChain";
