+++ src/components/EndorsementChain/EndorsementChainLayout/index.tsx
@@ -1,0 +1,1 @@
+export * from "./EndorsementChainLayout";
