+++ src/components/UI/SvgIcon/SvgIcon.tsx
@@ -194,6 +194,16 @@ export const SvgIconSave = () => {
   );
 };
 
+export const SvgIconAlert = () => {
+  return (
+    <g className="alert">
+      <path d="M10.29 3.86L1.82 18a2 2 0 0 0 1.71 3h16.94a2 2 0 0 0 1.71-3L13.71 3.86a2 2 0 0 0-3.42 0z" />
+      <polyline points="12 9 12 13" />
+      <polyline points="12 17 12.01 17" />
+    </g>
+  );
+};
+
 export const SvgIconTrash2 = () => {
   return (
     <g className="trash2">
